# docs/swagger.yaml
definitions:
  internal_auth.LoginPayload:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_auth.RefreshTokenPayload:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.RegisterPayload:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mobile_number:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_user.UpdateUser:
    properties:
      fname:
        type: string
      is_admin:
        type: boolean
      lname:
        type: string
      mobile_number:
        type: string
    type: object
  user-management_pkg_response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  user-management_pkg_response.PaginatedData:
    properties:
      items: {}
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a user management service and chat
  title: User Management API
  version: "1.0"
paths:
  /admin/api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users (admin only).
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/user-management_pkg_response.PaginatedData'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/user-management_pkg_response.PaginatedData'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.PaginatedData'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Admin
  /admin/api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user's profile by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a user profile by ID (admin only)
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update a user's profile by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateUser'
      produces:
      - application/json
      responses:
        "204":
          description: User updated successfully
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a user profile by ID (admin only)
      tags:
      - Admin
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Login using username and password
      parameters:
      - description: User login data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      summary: User login
      tags:
      - Authentication
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Provide a valid refresh token to get new access and refresh tokens
      parameters:
      - description: Refresh token data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "401":
          description: Unauthorized or expired token
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Register a new user by providing username, email, and password.
      parameters:
      - description: User registration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Validation errors
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      summary: Register a new user
      tags:
      - Authentication
  /api/v1/user:
    get:
      consumes:
      - application/json
      description: Get the profile of the authenticated user.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Validation errors
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates a user's profile using their ID and request body.
      parameters:
      - description: User update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateUser'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a user's profile by ID
      tags:
      - User
  /api/v1/user/logout:
    get:
      description: Logs out the currently authenticated user by removing their refresh
        token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/user-management_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"